Added a lot of more licensed spritesheets.

+++ src/engine/entities/npc/NPCCore.java
@@ -168,7 +168,7 @@ public class NPCCore extends Entity {
 
 	private void checkIfInWater(Tile tile) {
 		Tile leftCornerTile = level.getTile((int) (collissionBox.minX + position.x) >> bluePrint.atlas.sheet.getShiftOperator(), (int) (collissionBox.maxY + position.y) >> bluePrint.atlas.sheet.getShiftOperator());
-		Tile rightCornerTile = level.getTile((int) (collissionBox.maxX + position.x)>> bluePrint.atlas.sheet.getShiftOperator(), (int) (int) (collissionBox.maxY + position.y) >> bluePrint.atlas.sheet.getShiftOperator());
+		Tile rightCornerTile = level.getTile((int) (collissionBox.maxX + position.x)>> bluePrint.atlas.sheet.getShiftOperator(), (int) (collissionBox.maxY + position.y) >> bluePrint.atlas.sheet.getShiftOperator());
 		
 		if (leftCornerTile!= null && rightCornerTile != null) {
 			if (leftCornerTile.bluePrint.equals(Tiles.WATER_CLEAN) && rightCornerTile.bluePrint.equals(Tiles.WATER_CLEAN)) {
